rename `pool` to `buffer`
It isn't really a connection pool
